simplified gas objects, working toward full atmosphere fluxes and heating rates

+++ src/radiation.jl
@@ -6,32 +6,32 @@ export ν2f, f2ν, ν2λ, λ2ν, λ2f, f2λ
 """
 Convert wavenumber [cm``^{-1}``] to frequency [1/s]
 """
-ν2f(ν)::Float64 = 100.0*𝐜*ν
+ν2f(ν) = 100.0*𝐜*ν
 
 """
 Convert frequency [1/s] to wavenumber [cm``^{-1}``]
 """
-f2ν(f)::Float64 = f/(100.0*𝐜)
+f2ν(f) = f/(100.0*𝐜)
 
 """
 Convert wavenumber [cm``^{-1}``] to wavelength [m]
 """
-ν2λ(ν)::Float64 = 0.01/ν
+ν2λ(ν) = 0.01/ν
 
 """
 Convert wavelength [m] to wavenumber [cm``^{-1}``]
 """
-λ2ν(λ)::Float64 = 0.01/λ
+λ2ν(λ) = 0.01/λ
 
 """
 Convert wavelength [m] to frequency [1/s]
 """
-λ2f(λ)::Float64 = 𝐜/λ
+λ2f(λ) = 𝐜/λ
 
 """
 Convert frequency [1/s] to wavelength [m]
 """
-f2λ(f)::Float64 = f/𝐜
+f2λ(f) = f/𝐜
 
 #-------------------------------------------------------------------------------
 export planck, normplanck, stefanboltzmann, equilibriumtemperature
@@ -45,7 +45,7 @@ Compute black body intensity [W/m``^2``/cm``^{-1}``/sr] using [Planck's law](htt
 * `ν`: wavenumger [cm``^{-1}``]
 * `T`: temperature [Kelvin]
 """
-planck(ν, T)::Float64 = 100*2*𝐡*𝐜^2*(100*ν)^3/(exp(𝐡*𝐜*(100*ν)/(𝐤*T)) - 1)
+planck(ν, T) = 100*2*𝐡*𝐜^2*(100*ν)^3/(exp(𝐡*𝐜*(100*ν)/(𝐤*T)) - 1)
 
 """
     normplanck(ν, T)
@@ -62,14 +62,14 @@ yielding units of 1/cm``^{-1}``/sr.
 * `ν`: wavenumger [cm``^{-1}``]
 * `T`: temperature [Kelvin]
 """
-normplanck(ν, T)::Float64 = planck(ν, T)/stefanboltzmann(T)
+normplanck(ν, T) = planck(ν, T)/stefanboltzmann(T)
 
 """
     stefanboltzmann(T)
 
 Compute black body radiation power using the [Stefan-Boltzmann](https://en.wikipedia.org/wiki/Stefan%E2%80%93Boltzmann_law) law, ``σT^4`` [W/m``^2``].
 """
-stefanboltzmann(T)::Float64 = 𝛔*T^4
+stefanboltzmann(T) = 𝛔*T^4
 
 """
     equilibriumtemperature(F, A)
@@ -82,7 +82,7 @@ Compute the [planetary equilibrium temperature](https://en.wikipedia.org/wiki/Pl
 * `F`: stellar flux [W/m``^2``]
 * `A`: albedo
 """
-equilibriumtemperature(F, A)::Float64 = ((1 - A)*F/(4*𝛔))^(1/4)
+equilibriumtemperature(F, A) = ((1 - A)*F/(4*𝛔))^(1/4)
 
 """
     equilibriumtemperature(L, A, R)
@@ -96,7 +96,7 @@ Compute the [planetary equilibrium temperature](https://en.wikipedia.org/wiki/Pl
 * `A`: albedo
 * `R`: orbital distance [m]
 """
-equilibriumtemperature(L, A, R)::Float64 = (L*(1 - A)/(16*𝛔*π*R^2))^(1/4)
+equilibriumtemperature(L, A, R) = (L*(1 - A)/(16*𝛔*π*R^2))^(1/4)
 
 #-------------------------------------------------------------------------------
 export dτdP, transmittance, schwarzschild
@@ -115,14 +115,14 @@ where ``N_A`` is Avogadro's number.
 * `g`: gravitational acceleration [m/s``^2``]
 * `μ`: mean molar mass [kg/mole]
 """
-dτdP(σ, g, μ)::Float64 = 1e-4*σ*(𝐍𝐚/(μ*g))
+dτdP(σ, g, μ) = 1e-4*σ*(𝐍𝐚/(μ*g))
 
 """
     transmittance(τ)
 
 Evaluate transmittance from optical depth, ``t = e^{-τ}``
 """
-transmittance(τ)::Float64 = exp(-τ)
+transmittance(τ) = exp(-τ)
 
 """
     schwarzschild(I, ν, σ, T, P)
@@ -140,7 +140,7 @@ where ``B_ν`` is [`planck`](@ref)'s law.
 * `T`: temperature [K]
 * `P`: pressure [Pa]
 """
-schwarzschild(I, ν, σ, T, P)::Float64 = 1e-4*σ*(P/(𝐤*T))*(planck(ν,T) - I)
+schwarzschild(I, ν, σ, T, P) = 1e-4*σ*(P/(𝐤*T))*(planck(ν,T) - I)
 
 """
     schwarzschild(I, ν, σ, g, μ, T)
@@ -159,4 +159,4 @@ where ``B_ν`` is [`planck`](@ref)'s law and ``N_A`` is Avogadro's number.
 * `μ`: mean molar mass [kg/mole]
 * `T`: temperature [K]
 """
-schwarzschild(I, ν, σ, g, μ, T)::Float64 = 1e-4*σ*(𝐍𝐚/(μ*g))*(planck(ν,T) - I)
+schwarzschild(I, ν, σ, g, μ, T) = 1e-4*σ*(𝐍𝐚/(μ*g))*(planck(ν,T) - I)
